move get method tests into new test_data.py file
