+++ python/test/test_images.py
@@ -485,92 +485,6 @@ class TestImagesStats(PySparkTestCase):
         assert_true(array_equal(reduce(minimum, arys), minVal))
 
 
-class TestImagesGetters(PySparkTestCase):
-    def setUp(self):
-        super(TestImagesGetters, self).setUp()
-        self.ary1 = array([[1, 2], [3, 4]], dtype='int16')
-        self.ary2 = array([[5, 6], [7, 8]], dtype='int16')
-        self.images = ImagesLoader(self.sc).fromArrays([self.ary1, self.ary2])
-
-    def test_getMissing(self):
-        assert_is_none(self.images.get(-1))
-
-    def test_get(self):
-        assert_true(array_equal(self.ary2, self.images.get(1)))
-
-    def test_getAll(self):
-        vals = self.images.getAll([0, -1, 1, 0])
-        assert_equals(4, len(vals))
-        assert_true(array_equal(self.ary1, vals[0]))
-        assert_is_none(vals[1])
-        assert_true(array_equal(self.ary2, vals[2]))
-        assert_true(array_equal(self.ary1, vals[3]))
-
-    def test_getRanges(self):
-        vals = self.images.getRange(slice(None))
-        assert_equals(2, len(vals))
-        assert_equals(0, vals[0][0])
-        assert_equals(1, vals[1][0])
-        assert_true(array_equal(self.ary1, vals[0][1]))
-        assert_true(array_equal(self.ary2, vals[1][1]))
-
-        vals = self.images.getRange(slice(0, 1))
-        assert_equals(1, len(vals))
-        assert_equals(0, vals[0][0])
-        assert_true(array_equal(self.ary1, vals[0][1]))
-
-        vals = self.images.getRange(slice(1))
-        assert_equals(1, len(vals))
-        assert_equals(0, vals[0][0])
-        assert_true(array_equal(self.ary1, vals[0][1]))
-
-        vals = self.images.getRange(slice(1, 2))
-        assert_equals(1, len(vals))
-        assert_equals(1, vals[0][0])
-        assert_true(array_equal(self.ary2, vals[0][1]))
-
-        vals = self.images.getRange(slice(2, 3))
-        assert_equals(0, len(vals))
-
-        # raise exception if 'step' specified:
-        assert_raises(ValueError, self.images.getRange, slice(1, 2, 2))
-
-    def test_brackets(self):
-        vals = self.images[1]
-        assert_true(array_equal(self.ary2, vals))
-
-        vals = self.images[0:1]
-        assert_equals(1, len(vals))
-        assert_equals(0, vals[0][0])
-        assert_true(array_equal(self.ary1, vals[0][1]))
-
-        vals = self.images[:]
-        assert_equals(2, len(vals))
-        assert_equals(0, vals[0][0])
-        assert_equals(1, vals[1][0])
-        assert_true(array_equal(self.ary1, vals[0][1]))
-        assert_true(array_equal(self.ary2, vals[1][1]))
-
-        vals = self.images[1:4]
-        assert_equals(1, len(vals))
-        assert_equals(1, vals[0][0])
-        assert_true(array_equal(self.ary2, vals[0][1]))
-
-        vals = self.images[1:]
-        assert_equals(1, len(vals))
-        assert_equals(1, vals[0][0])
-        assert_true(array_equal(self.ary2, vals[0][1]))
-
-        vals = self.images[:1]
-        assert_equals(1, len(vals))
-        assert_equals(0, vals[0][0])
-        assert_true(array_equal(self.ary1, vals[0][1]))
-
-        assert_raises(KeyError, self.images.__getitem__, 2)  # equiv: self.images[2]
-
-        assert_raises(IndexError, self.images.__getitem__, slice(2, 3))  # equiv: self.images[2:3]
-
-
 class TestImagesUsingOutputDir(PySparkTestCaseWithOutputDir):
 
     @staticmethod
