+++ python/test/test_series.py
@@ -305,145 +305,3 @@ class TestSeriesMethods(PySparkTestCase):
 
         assert_raises(ValueError, setIndex, data, 5)
         assert_raises(ValueError, setIndex, data, [1, 2])
-
-
-class TestSeriesGetters(PySparkTestCase):
-    def setUp(self):
-        super(TestSeriesGetters, self).setUp()
-        self.dataLocal = [
-            ((0, 0), array([1.0, 2.0, 3.0], dtype='float32')),
-            ((0, 1), array([2.0, 2.0, 4.0], dtype='float32')),
-            ((1, 0), array([4.0, 2.0, 1.0], dtype='float32')),
-            ((1, 1), array([3.0, 1.0, 1.0], dtype='float32'))
-        ]
-        self.series = Series(self.sc.parallelize(self.dataLocal), dtype='float32', dims=(2, 2), index=[0, 1, 2])
-
-    def test_getMissing(self):
-        assert_is_none(self.series.get(-1))
-
-    def test_get(self):
-        expected = self.dataLocal[1][1]
-        assert_true(array_equal(expected, self.series.get((0, 1))))
-
-    def test_getAll(self):
-        vals = self.series.getAll([(0, 0), (17, 256), (1, 0), (0, 0)])
-        assert_equals(4, len(vals))
-        assert_true(array_equal(self.dataLocal[0][1], vals[0]))
-        assert_is_none(vals[1])
-        assert_true(array_equal(self.dataLocal[2][1], vals[2]))
-        assert_true(array_equal(self.dataLocal[0][1], vals[3]))
-
-    def test_getRanges(self):
-        vals = self.series.getRange([slice(2), slice(2)])
-        assert_equals(4, len(vals))
-        assert_equals(self.dataLocal[0][0], vals[0][0])
-        assert_equals(self.dataLocal[1][0], vals[1][0])
-        assert_equals(self.dataLocal[2][0], vals[2][0])
-        assert_equals(self.dataLocal[3][0], vals[3][0])
-        assert_true(array_equal(self.dataLocal[0][1], vals[0][1]))
-        assert_true(array_equal(self.dataLocal[1][1], vals[1][1]))
-        assert_true(array_equal(self.dataLocal[2][1], vals[2][1]))
-        assert_true(array_equal(self.dataLocal[3][1], vals[3][1]))
-
-        vals = self.series.getRange([slice(2), slice(1)])
-        assert_equals(2, len(vals))
-        assert_equals(self.dataLocal[0][0], vals[0][0])
-        assert_equals(self.dataLocal[2][0], vals[1][0])
-        assert_true(array_equal(self.dataLocal[0][1], vals[0][1]))
-        assert_true(array_equal(self.dataLocal[2][1], vals[1][1]))
-
-        vals = self.series.getRange([slice(None), slice(1, 2)])
-        assert_equals(2, len(vals))
-        assert_equals(self.dataLocal[1][0], vals[0][0])
-        assert_equals(self.dataLocal[3][0], vals[1][0])
-        assert_true(array_equal(self.dataLocal[1][1], vals[0][1]))
-        assert_true(array_equal(self.dataLocal[3][1], vals[1][1]))
-
-        vals = self.series.getRange([slice(None), slice(None)])
-        assert_equals(4, len(vals))
-        assert_equals(self.dataLocal[0][0], vals[0][0])
-        assert_equals(self.dataLocal[1][0], vals[1][0])
-        assert_equals(self.dataLocal[2][0], vals[2][0])
-        assert_equals(self.dataLocal[3][0], vals[3][0])
-        assert_true(array_equal(self.dataLocal[0][1], vals[0][1]))
-        assert_true(array_equal(self.dataLocal[1][1], vals[1][1]))
-        assert_true(array_equal(self.dataLocal[2][1], vals[2][1]))
-        assert_true(array_equal(self.dataLocal[3][1], vals[3][1]))
-
-        vals = self.series.getRange([0, slice(None)])
-        assert_equals(2, len(vals))
-        assert_equals(self.dataLocal[0][0], vals[0][0])
-        assert_equals(self.dataLocal[1][0], vals[1][0])
-        assert_true(array_equal(self.dataLocal[0][1], vals[0][1]))
-        assert_true(array_equal(self.dataLocal[1][1], vals[1][1]))
-
-        vals = self.series.getRange([0, 1])
-        assert_equals(1, len(vals))
-        assert_equals(self.dataLocal[1][0], vals[0][0])
-        assert_true(array_equal(self.dataLocal[1][1], vals[0][1]))
-
-        vals = self.series.getRange([slice(2, 3), slice(None)])
-        assert_equals(0, len(vals))
-
-        # raise exception if 'step' specified:
-        assert_raises(ValueError, self.series.getRange, [slice(0, 4, 2), slice(2, 3)])
-
-    def test_brackets(self):
-        # returns just value; calls `get`
-        vals = self.series[(1, 0)]
-        assert_true(array_equal(self.dataLocal[2][1], vals))
-
-        # tuple isn't needed; returns just value, calls `get`
-        vals = self.series[0, 1]
-        assert_true(array_equal(self.dataLocal[1][1], vals))
-
-        # if slices are passed, calls `getRange`, returns keys and values
-        vals = self.series[0:1, 1:2]
-        assert_equals(1, len(vals))
-        assert_equals(self.dataLocal[1][0], vals[0][0])
-        assert_true(array_equal(self.dataLocal[1][1], vals[0][1]))
-
-        # if slice extends out of bounds, return only the elements that are in bounds
-        vals = self.series[:4, :1]
-        assert_equals(2, len(vals))
-        assert_equals(self.dataLocal[0][0], vals[0][0])
-        assert_equals(self.dataLocal[2][0], vals[1][0])
-        assert_true(array_equal(self.dataLocal[0][1], vals[0][1]))
-        assert_true(array_equal(self.dataLocal[2][1], vals[1][1]))
-
-        # empty slice works
-        vals = self.series[:, 1:2]
-        assert_equals(2, len(vals))
-        assert_equals(self.dataLocal[1][0], vals[0][0])
-        assert_equals(self.dataLocal[3][0], vals[1][0])
-        assert_true(array_equal(self.dataLocal[1][1], vals[0][1]))
-        assert_true(array_equal(self.dataLocal[3][1], vals[1][1]))
-
-        # multiple empty slices work
-        vals = self.series[:, :]
-        assert_equals(4, len(vals))
-        assert_equals(self.dataLocal[0][0], vals[0][0])
-        assert_equals(self.dataLocal[1][0], vals[1][0])
-        assert_equals(self.dataLocal[2][0], vals[2][0])
-        assert_equals(self.dataLocal[3][0], vals[3][0])
-        assert_true(array_equal(self.dataLocal[0][1], vals[0][1]))
-        assert_true(array_equal(self.dataLocal[1][1], vals[1][1]))
-        assert_true(array_equal(self.dataLocal[2][1], vals[2][1]))
-        assert_true(array_equal(self.dataLocal[3][1], vals[3][1]))
-
-        # mixing slices and individual indicies works:
-        vals = self.series[0, :]
-        assert_equals(2, len(vals))
-        assert_equals(self.dataLocal[0][0], vals[0][0])
-        assert_equals(self.dataLocal[1][0], vals[1][0])
-        assert_true(array_equal(self.dataLocal[0][1], vals[0][1]))
-        assert_true(array_equal(self.dataLocal[1][1], vals[1][1]))
-
-        # trying to getitem a key that doesn't exist raises KeyError
-        # this differs from `get` behavior but is consistent with python dict
-        # see object.__getitem__ in https://docs.python.org/2/reference/datamodel.html
-        assert_raises(KeyError, self.series.__getitem__, (25, 17))  # equiv: self.series[(25, 17)]
-
-        # passing a range that is completely out of bounds throws IndexError
-        # note that if a range is only partly out of bounds, it will return what elements the slice does include
-        assert_raises(IndexError, self.series.__getitem__, [slice(2, 3), slice(None)])  # series[2:3,:]
